refactor(example): split wallet standard features into components
